refactor(driver): Refactor driver entities
- Refactored driver entities folder
- Refactor driver public api's
