+++ packages/client/__tests__/e2e/app/drivers/page.test.ts
@@ -2,7 +2,7 @@ import { expect } from '@playwright/test';
 import { test } from '~tests-utils/e2e/server/MockApiTest';
 import { setupServer, closeServer } from '~tests-utils/e2e/server/MockFastifyServer';
 
-import { URLS } from '~entities/driver/api/urls';
+import { DRIVER_URLS } from '~entities/driver/api';
 import { CONSTRUCTOR_URLS } from '~entities/constructor/api';
 
 import { DriversMock } from '~mocks/entities/driver/Driver.mock';
@@ -15,7 +15,7 @@ test.afterEach(async ({ server }) => {
 
 test('render drivers list', async ({ page, server }) => {
     await setupServer(server, {
-        url: URLS.index,
+        url: DRIVER_URLS.index,
         method: 'GET',
         handler: function (_, reply) {
             reply.send({
@@ -43,7 +43,7 @@ test("should go to driver page after driver's name click", async ({ page, server
     await setupServer(
         server,
         {
-            url: URLS.index,
+            url: DRIVER_URLS.index,
             method: 'GET',
             handler: function (_, reply) {
                 reply.send({
@@ -53,7 +53,7 @@ test("should go to driver page after driver's name click", async ({ page, server
             },
         },
         {
-            url: URLS.ref(driver.ref),
+            url: DRIVER_URLS.ref(driver.ref),
             method: 'GET',
             handler: function (_, reply) {
                 reply.send(driver);
@@ -80,7 +80,7 @@ test("should go to constructor page after constructor's name click", async ({ pa
     await setupServer(
         server,
         {
-            url: URLS.index,
+            url: DRIVER_URLS.index,
             method: 'GET',
             handler: function (_, reply) {
                 reply.send({
@@ -107,7 +107,7 @@ test("should go to constructor page after constructor's name click", async ({ pa
 
 test('should render breadcrumbs correctly', async ({ page, server }) => {
     await setupServer(server, {
-        url: URLS.index,
+        url: DRIVER_URLS.index,
         method: 'GET',
         handler: function (_, reply) {
             reply.send({
