+++ packages/client/src/entities/driver/api/index.ts
@@ -1,2 +1,3 @@
-export * as DRIVER_URLS from './urls';
+export { URLS as DRIVER_URLS } from './urls';
+export { useDriverServer } from './useDriverServer';
 export { useDriversServer } from './useDriversServer';
