+++ packages/client/src/entities/driver/api/useDriverServer.ts
@@ -1,10 +1,10 @@
 import { axios } from '~shared/api/axios';
-import { URLS } from '~entities/driver/api/urls';
-import { DriverType } from '~entities/driver/type';
+import { DRIVER_URLS } from '~entities/driver/api';
+import type { DriverType } from '~entities/driver';
 
 export async function useDriverServer(ref: string) {
     try {
-        const { data: driver } = await axios.get<DriverType>(URLS.ref(ref));
+        const { data: driver } = await axios.get<DriverType>(DRIVER_URLS.ref(ref));
 
         return { driver };
     } catch (error) {
