+++ packages/client/src/entities/driver/index.ts
@@ -1,2 +1,1 @@
 export type { DriverType } from './type';
-export * as DriverApi from './api';
