+++ packages/client/src/entities/driver/ui/index.ts
@@ -1,0 +1,1 @@
+export { DriverListCard } from './list-card';
